rename utils/config.ts
